requirements-pipeline.txt
numpy==2.1.3
xarray==2025.1.2
zarr==2.18.3
s3fs==2025.2.0
numcodecs==0.15.1
pillow==11.1.0
pyproj==3.7.1
rasterio==1.4.3
